[objc-appscript] updated project (10.6+, no GC); fixed misc bugs, warnings

+++ objc-appscript/trunk/src/Appscript/parser.h
@@ -14,7 +14,7 @@
 @interface ASParserDef : NSObject {
 	NSString *name;
 	OSType code;
-	unsigned hash;
+	NSUInteger hash;
 }
 
 - (id)initWithName:(NSString*)name_ code:(OSType)code_;
